security layers on endpoints

+++ src/main/java/com/empSystem/services/DepartmentServiceImpl.java
@@ -4,8 +4,11 @@ import com.empSystem.abstracts.DepartmentService;
 import com.empSystem.dtos.DepartmentCreate;
 import com.empSystem.dtos.DepartmentUpdate;
 import com.empSystem.entities.Department;
+import com.empSystem.entities.Employee;
+import com.empSystem.exceptions.BadRequestException;
 import com.empSystem.exceptions.NotFoundException;
 import com.empSystem.repository.DepartmentRepo;
+import com.empSystem.repository.EmployeeRepo;
 import com.empSystem.utils.LocaleMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +23,9 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Autowired
     private LocaleMessage localeMessage;
+
+    @Autowired
+    private EmployeeRepo employeeRepo;
 
     @Override
     public List<Department> getAll() {
@@ -51,9 +57,20 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public void deleteOne(UUID id) {
+        List<Employee> employeesInThisDepartment = employeeRepo.findAllByDepartmentId(id);
+        if (!employeesInThisDepartment.isEmpty()) {
+            throw new BadRequestException("Can't delete department with employees. Reassign or delete employees first");
+        }
+
         Department department = departmentRepo.findById(id).orElseThrow(() ->
                 new NotFoundException(localeMessage.getMessage("department.not.found", id.toString())));
 
+
         departmentRepo.delete(department);
+    }
+
+    @Override
+    public Department findByName(String name) {
+        return departmentRepo.findByName(name).orElseThrow(() -> new NotFoundException("Not Found Department under name: " + name));
     }
 }
